add FlowHistory
