+++ mobile-ui/src/components/flow/components/FlowHistory.tsx
@@ -3,7 +3,7 @@ import {FlowViewReactContext} from "@/components/flow/view";
 import Descriptions from "@/components/descriptions";
 import {FormField} from "@/components/form/types";
 import moment from "moment";
-import {Divider, Tag} from "antd-mobile";
+import {Tag} from "antd-mobile";
 
 const fields = [
     {
@@ -134,8 +134,6 @@ const FlowHistory = () => {
     const flowViewReactContext = useContext(FlowViewReactContext);
     const flowRecordContext = flowViewReactContext?.flowRecordContext;
 
-    const historyRecords = flowRecordContext?.getHistoryRecords();
-
     const currentFlowRecord = flowRecordContext?.getCurrentFlowRecord();
 
     return (
@@ -158,30 +156,6 @@ const FlowHistory = () => {
                         };
                     }}
                 />
-            </div>
-            <Divider>流转历史</Divider>
-            <div className={"flow-history-list"}>
-                {historyRecords && historyRecords.map((historyRecord: any, index: number) => {
-                    const nodeName = flowRecordContext?.getNode(historyRecord.nodeCode)?.name;
-                    const recodeType = recodeTypeConvert(historyRecord);
-                    const createName = historyRecord.createOperator?.name;
-                    const createTime = moment(historyRecord.createTime).format('YYYY-MM-DD HH:mm:ss');
-                    const flowAdvice = historyRecord.opinion?.advice || '暂无意见';
-                    return (
-                        <div className={"flow-history-list-item"}>
-                            <div className={"flow-history-list-item-line"}>
-                                {historyRecords.length > (index + 1) &&  <div className={"flow-history-list-item-dot"}/>}
-                            </div>
-                            <div className={"flow-history-list-item-content"}>
-                                <div className={"flow-history-list-item-title"}>{nodeName}</div>
-                                <div className={"flow-history-list-item-attr"}>状态:{recodeType}</div>
-                                <div className={"flow-history-list-item-attr"}>创建人:{createName}</div>
-                                <div className={"flow-history-list-item-attr"}>创建事件:{createTime}</div>
-                                <div className={"flow-history-list-item-attr"}>审批意见:{flowAdvice}</div>
-                            </div>
-                        </div>
-                    )
-                })}
             </div>
         </div>
     )
